Adjusted data models and system: - move mutex into entries ("line locks") for better granularity - Adjusted parameters and return types for methods in `Account`, `BankingCore`, `DataAdapter` and `Database` - Added constraint contracts (lambdas and exceptions) for future usage
Added skeletons for parsing utils, payload shape, as well as `EntryLog` and `Entry` for database implementation

Added some first test implementations

+++ src/main/kotlin/client/Account.kt
@@ -6,10 +6,10 @@ class Account {
 
   // kata interface
 
-  fun deposit(value: Int, timestamp: Date): Unit =
+  fun deposit(value: Double, timestamp: Date): Unit =
     TODO("Not implemented")
 
-  fun withdraw(value: Int, timestamp: Date): Unit =
+  fun withdraw(value: Double, timestamp: Date): Unit =
     TODO("Not implemented")
 
   fun printStatement(): String =
